# Conflicts:
#	src/main/java/tn/esprit/gestionbancaire/services/MailService.java
#	src/main/java/tn/esprit/gestionbancaire/services/impl/MailServiceImpl.java
#	src/main/java/tn/esprit/gestionbancaire/utils/MailTemplates.java

+++ src/main/java/tn/esprit/gestionbancaire/services/impl/AdministrativeDocumentServiceImpl.java
@@ -3,13 +3,18 @@ package tn.esprit.gestionbancaire.services.impl;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import tn.esprit.gestionbancaire.enums.CreditStatus;
 import tn.esprit.gestionbancaire.exception.ErrorCodes;
 import tn.esprit.gestionbancaire.exception.InvalidEntityException;
+import tn.esprit.gestionbancaire.exception.InvalidOperationException;
 import tn.esprit.gestionbancaire.model.AdministrativeDocument;
+import tn.esprit.gestionbancaire.model.Credit;
 import tn.esprit.gestionbancaire.repository.AdministrativeDocumentRepository;
 import tn.esprit.gestionbancaire.services.AdministrativeDocumentService;
+import tn.esprit.gestionbancaire.services.CreditService;
 import tn.esprit.gestionbancaire.validator.AdministrativeDocumentValidator;
 
+import java.time.Instant;
 import java.util.List;
 import java.util.Optional;
 
@@ -18,28 +23,43 @@ import java.util.Optional;
 public class AdministrativeDocumentServiceImpl implements AdministrativeDocumentService {
 
     private AdministrativeDocumentRepository administrativeDocumentRepository;
+    private CreditService creditService;
 
     @Autowired
-    public AdministrativeDocumentServiceImpl(AdministrativeDocumentRepository administrativeDocumentRepository) {
+    public AdministrativeDocumentServiceImpl(AdministrativeDocumentRepository administrativeDocumentRepository,CreditService creditService) {
         this.administrativeDocumentRepository = administrativeDocumentRepository;
+        this.creditService = creditService;
 
     }
 
 
     @Override
     public AdministrativeDocument save(AdministrativeDocument administrativeDocument) {
+
+       // Credit credit = administrativeDocument.getCredit();
         List<String> errors = AdministrativeDocumentValidator.validate(administrativeDocument);
         if (!errors.isEmpty()) {
             log.error("administrativeDocument is not valid {}", administrativeDocument);
             throw new InvalidEntityException("Credit is not valid", ErrorCodes.CREDIT_NOT_VALID, errors);
         }
-
+        Credit credit = creditService.findById(administrativeDocument.getCredit().getId());
+        List<String> notes = credit.getNotes();
+        notes.add("Add New " + administrativeDocument.getAdministrativeDocumentType() + " Document has been added");
+        credit.setNotes(notes);
+        administrativeDocument.setCreationDate(Instant.now());
+        administrativeDocument.setLastModifiedDate(Instant.now());
+        credit.setLastModifiedDate(Instant.now());
         return administrativeDocumentRepository.save(administrativeDocument);
     }
 
     @Override
     public AdministrativeDocument update(AdministrativeDocument administrativeDocument) {
-        return null;
+        Credit credit = creditService.findById(administrativeDocument.getCredit().getId());
+        List<String> notes = credit.getNotes();
+        notes.add(administrativeDocument.getAdministrativeDocumentType() + "Document has been updated");
+        credit.setNotes(notes);
+        administrativeDocument.setLastModifiedDate(Instant.now());
+        return administrativeDocumentRepository.save(administrativeDocument);
     }
 
     @Override
@@ -59,7 +79,15 @@ public class AdministrativeDocumentServiceImpl implements AdministrativeDocument
             log.error("Administrative Document ID is null");
             return;
         }
-        administrativeDocumentRepository.deleteById(id);
+        Optional<AdministrativeDocument> administrativeDocument = administrativeDocumentRepository.findById(id);
+
+        if (administrativeDocument.isPresent()) {
+            administrativeDocument.get().setCredit(null);
+            administrativeDocumentRepository.save(administrativeDocument.get());
+            administrativeDocumentRepository.deleteById(id);
+        }
+
+
     }
 
     @Override
@@ -67,17 +95,18 @@ public class AdministrativeDocumentServiceImpl implements AdministrativeDocument
         if (id == null) {
             log.error("Administrative Document ID is null");
         }
-
         Optional<AdministrativeDocument> byId = administrativeDocumentRepository.findById(id);
         AdministrativeDocument administrativeDocument = null;
         if (byId.isPresent()) {
             administrativeDocument = byId.get();
             administrativeDocument.setPhoto(imageURL);
             administrativeDocumentRepository.save(administrativeDocument);
+            Credit credit = creditService.findById(administrativeDocument.getCredit().getId());
+            List<String> notes = credit.getNotes();
+            notes.add("Add New image has been add to " + administrativeDocument.getAdministrativeDocumentType() + "Document N:" + id);
+            credit.setNotes(notes);
+            administrativeDocument.setLastModifiedDate(Instant.now());
         }
-
         return administrativeDocument;
-
-
     }
 }
